Extract: routes and controller logic into /backend
* Update unit tests paths
* Rename folders and files for clarity according to use
* All logic in controller
