+++ index.js
@@ -13,7 +13,7 @@ const app = express();
  * Model route naming
  */ 
 
-const Bookshop = require("./models/bookshop");
+// const Bookshop = require("./backend/models/bookshop");
 
 /**
  * MIDDLEWARE
@@ -27,22 +27,15 @@ app.use(express.urlencoded({ extended: false }))
 /**
  * ROUTE SETUP
  */
+// router attribution
+
+const BookShopRouter = require("./backend/routes/bookshop")
 
 app.get("/", (req, res) => {
   res.redirect("/openbook")
 })
 
-app.get("/openbook", (req, res) => {
-  const bookshopIds = []
-  Bookshop.find((err, bookshops) => {
-    bookshops.forEach((bookshop) => {
-      let shop = bookshop.name.replace(/'/g, "").split(" ")
-      shop = shop.join("").toLowerCase()
-      bookshopIds.push(shop)
-    })
-    res.render('index', { bookshops: bookshops, bookshopIds: bookshopIds })
-  })
-});
+app.use("/openbook", BookShopRouter)
 
 app.get("/users/new", (req, res) => {
   res.render('sign_up')
